feat/ add components and logic for steps 1-3

+++ src/utils/api.js
@@ -1,3 +1,5 @@
+import { getToken } from "./auth";
+
 const baseUrl = "http://localhost:3001";
 
 function verifyResponse(res) {
@@ -9,9 +11,18 @@ function getItems() {
 }
 
 function addItem({ name, imageUrl, weather }) {
+  const token = getToken();
+
+  if (!token) {
+    return Promise.reject({
+      error: "You must be logged in to add items",
+    });
+  }
+
   return fetch(`${baseUrl}/items`, {
     method: "POST",
     headers: {
+      authorization: `Bearer ${token}`,
       "Content-Type": "application/json",
     },
     body: JSON.stringify({ name, imageUrl, weather }),
@@ -19,9 +30,46 @@ function addItem({ name, imageUrl, weather }) {
 }
 
 function deleteItem(id) {
+  const token = getToken();
+
+  if (!token) {
+    return Promise.reject({
+      error: "You must be logged in to delete items",
+    });
+  }
+
   return fetch(`${baseUrl}/items/${id}`, {
     method: "DELETE",
+    headers: {
+      authorization: `Bearer ${token}`,
+    },
   }).then(verifyResponse);
 }
 
-export { getItems, addItem, deleteItem, verifyResponse, baseUrl };
+function updateUserInfo({ name, avatar }) {
+  const token = getToken();
+
+  if (!token) {
+    return Promise.reject({
+      error: "You must be logged in to update your profile",
+    });
+  }
+
+  return fetch(`${baseUrl}/users/me`, {
+    method: "PATCH",
+    headers: {
+      authorization: `Bearer ${token}`,
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify({ name, avatar }),
+  }).then(verifyResponse);
+}
+
+export {
+  getItems,
+  addItem,
+  deleteItem,
+  verifyResponse,
+  updateUserInfo,
+  baseUrl,
+};
